Commit 1.4b - Starting receipt printing
So far I am testing the receipt printing on a SwingWorker object. Right now the receipt only displays the entire arraylist of newTransaction. Will work on this more tomorrow.

+++ src/com/stockItems.java
@@ -4,7 +4,7 @@ public class stockItems {
     private String name;
     private float price;
     private int amount;
-    private double barcode;
+    private long barcode;
     private int plu;
     private int active;
 
@@ -33,10 +33,10 @@ public class stockItems {
         this.amount = amount;
     }
 
-    public double getBarcode() {
+    public long getBarcode() {
         return barcode; };
 
-    public void setBarcode(double barcode){
+    public void setBarcode(long barcode){
         this.barcode = barcode;
     }
